The predlast update before final

+++ Assets/Scripts/Projectiles/Projectile.cs
@@ -6,7 +6,7 @@ public abstract class Projectile : MonoBehaviour
 {
     public MoveableEnemy Enemy { get; set; }
     public MoveableEnemy CurrentEnemy { get; set; }
-    protected int Damage { get; set; }
+    protected float Damage { get; set; }
 
     public void SetEnemy(MoveableEnemy enemy)
     {
